% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getLam_olasso}
\alias{getLam_olasso}
\title{Get the two (theoretical) values of lambdas used in the organic lasso}
\usage{
getLam_olasso(x)
}
\arguments{
\item{x}{design matrix}
}
\description{
Get the two (theoretical) values of lambdas used in the organic lasso
}
