% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getLam_slasso}
\alias{getLam_slasso}
\title{Get the two (theoretical) values of lambdas used in scaled lasso}
\usage{
getLam_slasso(n, p)
}
\arguments{
\item{n}{number of observations}

\item{p}{number of features}
}
\description{
Get the two (theoretical) values of lambdas used in scaled lasso
}
