% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-natural.R
\docType{package}
\name{natural}
\alias{natural}
\alias{natural-package}
\title{natural: Natural and Organic lasso estimates of error variance in high-dimensional linear models}
\description{
The package contains implementation of the two methods introduced in Yu, Bien (2017) \url{https://arxiv.org/abs/1712.02412}.
}
\details{
The main functions are \code{\link{nlasso_cv}}, \code{\link{olasso_cv}}, and \code{\link{olasso}}.
}
