% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olasso_api.R
\name{olasso_slow}
\alias{olasso_slow}
\title{Solve organic lasso problem with a single value of lambda
The lambda values are for slow rates, which could give less satisfying results}
\usage{
olasso_slow(x, y, thresh = 1e-08)
}
\arguments{
\item{x}{An \code{n} by \code{p} design matrix. Each row is an observation of \code{p} features.}

\item{y}{A response vector of size \code{n}.}

\item{thresh}{Threshold value for underlying optimization algorithm to claim convergence. Default to be \code{1e-8}.}
}
\description{
Solve organic lasso problem with a single value of lambda
The lambda values are for slow rates, which could give less satisfying results
}
