% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_cv.R
\name{plot.natural.cv}
\alias{plot.natural.cv}
\title{plot a natural.cv object}
\usage{
\method{plot}{natural.cv}(x, ...)
}
\arguments{
\item{x}{an object of class \code{natural.cv}, as returned by
\code{\link{nlasso_cv}} and \code{\link{olasso_cv}}}

\item{...}{additional argument(not used here, only for S3 generic/method consistency)}
}
\description{
This function is adapted from the \pkg{ggb} R package.
}
