% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_path.R
\name{print.natural.path}
\alias{print.natural.path}
\title{print a natural.path object}
\usage{
\method{print}{natural.path}(x, ...)
}
\arguments{
\item{x}{an object of class \code{natural.path}, as returned by
\code{\link{nlasso_path}} and \code{\link{olasso_path}}}

\item{...}{additional argument(not used here, only for S3 generic/method consistency)}
}
\description{
This function is adapted from the \pkg{ggb} R package.
}
