% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardize}
\alias{standardize}
\title{Standardize the n -by- p design matrix X to have
column means zero and ||X_j||_2^2 = n for all j}
\usage{
standardize(x, center = TRUE)
}
\arguments{
\item{x}{design matrix}

\item{center}{should we set column means equal to zero}
}
\description{
Standardize the n -by- p design matrix X to have
column means zero and ||X_j||_2^2 = n for all j
}
